Add native macOS (Apple Silicon) build support (#4)
* Port to native macOS (Apple Silicon)

Adds an APPLE branch to CMakeLists (sdl12-compat, keg-only openssl@3
resolved via brew --prefix, -framework OpenGL) plus MenuBarHelper.m to
supply NSMenu_setMenuBarVisible without SDLMain.m's duplicate main():
under __mac__ the app's main becomes SDL_main, the real entry point
comes from sdl12-compat's libSDLmain.a, and compiling all of SDLMain.m
would define main twice.

gameSDL.cpp disables sdl12-compat's GL-scaling FBO on mac, which went
incomplete on window-server events (occlusion, display changes) and
froze the game mid-session with GL_INVALID_FRAMEBUFFER_OPERATION.
setenv uses overwrite=0 so SDL12COMPAT_OPENGL_SCALING=1 in the
environment can still force scaling back on.

hetuwmod.cpp gains an <sstream> include that libc++ requires for its
std::stringstream uses (libstdc++ pulls it in transitively).

README gains a macOS section: Homebrew dependencies, fetching the two
vendored library headers with curl (the setup script uses wget, which
stock macOS lacks), building, and running against an existing mac OHOL
data folder.

Linux and Windows builds are unaffected: the CMake changes live inside
a new elseif(APPLE) branch and the gameSDL.cpp change is #ifdef __mac__
only. This is source-only support; no .app bundle, signing, or release
artifacts.

* Fix clang narrowing errors in chat-log brace initializers

xd/yd are ints; brace-initializing a doublePair from them is an
int-to-double narrowing conversion. GCC accepts it (at most a
-Wnarrowing diagnostic) but clang rejects it outright:

  error: non-constant-expression cannot be narrowed from type 'int'
  to 'double' in initializer list [-Wc++11-narrowing]

Cast explicitly; the values are exact in double, so behavior is
unchanged. This unbreaks any clang build, not just macOS.

* Add a macOS build-check CI job

Build-only on the free macos-latest (arm64) runners: brew deps, the two
vendored headers, cmake build. No release artifacts; it exists so mac
support stays verifiably green without anyone needing a Mac on hand.

+++ gameSource/LivingLifePage.cpp
@@ -21283,7 +21283,7 @@ void LivingLifePage::step() {
                             if( firstSpace != NULL ) {
 
                                 // YummyLife: chat scrollback
-                                doublePair player_pos = { existing->xd, existing->yd };
+                                doublePair player_pos = { (double)existing->xd, (double)existing->yd };
                                 time_t timestamp = time(NULL);
                                 YummyLife::ChatLog::add(
                                     existing->name,
